Version 1.1.0: Support for Mode.OFF + Mode.DISABLED.
Mode.DISABLED replaces Mode.OFF, i.e. it disables the capturing mechanism but it is possible to enable it at runtime. The proxy is is still weaved into the methods that are annotated with @Capturable. Mode.OFF prevents that the advice is registered and makes it impossible to enable the mechanism at runtime. This can be used for production systems where performance matters.